Docker build appears to work
I saw the build_engine/cloudbuild.yaml build get all the way to uploading the image (and then hang), so I believe the build works.

This removes all the ssh stuff, I thought it would be needed for gclient, but it turns out it was just that I had accidentally specified an ssh checkout for the root (which we don't need).