Make build.sh work locally
build.sh wasn't tested before this, but I've now validated that it works locally with an (optional) 3rd arg which points to the engine location (instead of /engine).

This unfortunately also includes a change that won't work inside docker of depending on an NDK outside of Flutter's since (at least my copy) of the NDK Flutter engine pulled down does not seem to include all of the necessary arm64 dependencies.

https://github.com/flutter/flutter/issues/124280