Remove updater hacks and add patch upload
Remove the updater copy hacks, those are no longer needed now that the engine pulls down uploader itself (patch incoming to fix the engine build.gn files).

Also added a hack to zip/upload the patch tool for darwin-arm64.  We need a better solution for that (which may involve just re-writing it in Dart?)